Makes navbar a container and import actions
